Tailored project structure.
* no cmd, no pkg directory.
* internal now have transport for handler and common for utility methods.
* main.go is now responsible to run all servers and load balancers.
