+++ internal/common/env_vars_test.go
@@ -1,4 +1,4 @@
-package utils
+package common
 
 import (
 	"os"
