+++ internal/common/srvidgen_test.go
@@ -1,4 +1,4 @@
-package srvidgen
+package common
 
 import (
 	"testing"
